DAH-Capstone/Flutter_Project/Capstone 2/Dar Al-Hekma backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==25.3.0
backcall==0.2.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.1.3
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
executing==1.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.23.3
ipython==8.14.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.55
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.33
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.4.2
numpy==2.2.5
openai==1.75.0
orjson==3.10.16
packaging==24.2
parso==0.8.3
pickleshare==0.7.5
pillow==11.2.1
prompt-toolkit==3.0.38
propcache==0.3.1
psutil==5.9.5
pure-eval==0.2.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.15.1
pymongo==4.12.0
pypdf==5.4.0
python-dotenv==1.1.0
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.2
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tornado==6.3.2
tqdm==4.67.1
traitlets==5.9.0
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.6
Werkzeug==3.1.3
yarl==1.20.0
zstandard==0.23.0
